Update security
As soon as the trigPin sends a signal, the echoPin will go HIGH. That will trigger the pulseIn() function to start a timer. When the reflected pulse comes back to the receiver, the echoPin will drop back to LOW, which will cause the pulseIn() function to stop the timer. Knowing the value of the speed of sound (dry air @ 20 degrees C = 343m/s) we can find out the distance.